requirements.txt
numpy
pandas
scipy
opencv-python-headless
Pillow
matplotlib
scikit-learn
torch
torchvision
streamlit
fastapi
uvicorn
python-multipart
redis
requests
pytest
python-dotenv
tqdm
click
pydantic
PyYAML
